api: split into router/error/customer_ws/handlers/sse modules
Line-for-line moves: mod.rs keeps AppState/Tenancy, the router, serve/drain and auth;
customer_ws (WebSocket gateway), error (Failure + map_err), sessions, sandbox_files,
children, storage and sse each hold one concern. api.rs 2,531 -> 8 files, largest 426
lines. The api_code literal test now scans all of them.
